Tweak for very short latencies and the CDF algorithm for the server distribution.

+++ src/main/java/com/allanbank/mongodb/connection/state/ServerState.java
@@ -250,7 +250,7 @@ public class ServerState {
      *            The latency seen sending a request and receiving a reply from
      *            the server.
      */
-    public void updateAverageLatency(final long latency) {
+    public void updateAverageLatency(final double latency) {
         final double oldAverage = Double.longBitsToDouble(myAverageLatency
                 .get());
         if (Double.MAX_VALUE == oldAverage) {
